Add solution for issue classification and problem solving
This commit adds the solution for issue classification and problem solvi
ng in the `solve_issue` method of the `GithubTool` class. It utilizes
Op
enAI function calling (GPT-4) to generate solutions for the provided
`is
sue_message`. The `IssueSolution` class is also introduced, which
employ
s a `RunnableBranch` for problem solving as per the use case.